preparation before cobra-cli execution & run cobra-cli
