+++ agent/agent_test.go
@@ -1,4 +1,4 @@
-package dagu
+package agent
 
 import (
 	"net/http"
