+++ cmd/dry_tmp.go
@@ -1,0 +1,47 @@
+package cmd
+
+import (
+	"os"
+
+	"github.com/urfave/cli/v2"
+	"github.com/yohamta/dagu/agent"
+	"github.com/yohamta/dagu/internal/config"
+	"github.com/yohamta/dagu/internal/utils"
+)
+
+func newDryCommand() *cli.Command {
+	cl := &config.Loader{
+		HomeDir: utils.MustGetUserHomeDir(),
+	}
+	return &cli.Command{
+		Name:  "dry",
+		Usage: "dagu dry [--params=\"<params>\"] <config>",
+		Flags: []cli.Flag{
+			&cli.StringFlag{
+				Name:     "params",
+				Usage:    "parameters",
+				Value:    "",
+				Required: false,
+			},
+		},
+		Action: func(c *cli.Context) error {
+			config_file_path := c.Args().Get(0)
+			cfg, err := cl.Load(config_file_path, c.String("params"))
+			if err != nil {
+				return err
+			}
+			return dryRun(cfg)
+		},
+	}
+}
+
+func dryRun(cfg *config.Config) error {
+	a := &agent.Agent{AgentConfig: &agent.AgentConfig{
+		DAG: cfg,
+		Dry: true,
+	}}
+	listenSignals(func(sig os.Signal) {
+		a.Signal(sig)
+	})
+	return a.Run()
+}
