+++ cmd/server.go
@@ -1,40 +1,40 @@
-package main
+/*
+Copyright © 2022 NAME HERE <EMAIL ADDRESS>
+
+*/
+package cmd
 
 import (
-	"os"
-	"path"
+	"fmt"
 
-	"github.com/urfave/cli/v2"
-	"github.com/yohamta/dagu/internal/admin"
-	"github.com/yohamta/dagu/internal/utils"
+	"github.com/spf13/cobra"
 )
 
-func newServerCommand() *cli.Command {
-	l := &admin.Loader{}
-	return &cli.Command{
-		Name:  "server",
-		Usage: "dagu server",
-		Action: func(c *cli.Context) error {
-			cfg, err := l.LoadAdminConfig(
-				path.Join(utils.MustGetUserHomeDir(), ".dagu/admin.yaml"))
-			if err == admin.ErrConfigNotFound {
-				cfg = admin.DefaultConfig()
-			} else if err != nil {
-				return err
-			}
-			return startServer(cfg)
-		},
-	}
+// serverCmd represents the server command
+var serverCmd = &cobra.Command{
+	Use:   "server",
+	Short: "A brief description of your command",
+	Long: `A longer description that spans multiple lines and likely contains examples
+and usage of using your command. For example:
+
+Cobra is a CLI library for Go that empowers applications.
+This application is a tool to generate the needed files
+to quickly create a Cobra application.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("server called")
+	},
 }
 
-func startServer(cfg *admin.Config) error {
-	server := admin.NewServer(cfg)
+func init() {
+	rootCmd.AddCommand(serverCmd)
 
-	listenSignals(func(sig os.Signal) {
-		server.Shutdown()
-	})
+	// Here you will define your flags and configuration settings.
 
-	err := server.Serve()
-	utils.LogErr("running server", err)
-	return err
+	// Cobra supports Persistent Flags which will work for this command
+	// and all subcommands, e.g.:
+	// serverCmd.PersistentFlags().String("foo", "", "A help for foo")
+
+	// Cobra supports local flags which will only run when this command
+	// is called directly, e.g.:
+	// serverCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
