+++ cmd/start.go
@@ -1,49 +1,40 @@
-package main
+/*
+Copyright © 2022 NAME HERE <EMAIL ADDRESS>
+
+*/
+package cmd
 
 import (
-	"os"
+	"fmt"
 
-	"github.com/urfave/cli/v2"
-	"github.com/yohamta/dagu"
-	"github.com/yohamta/dagu/internal/config"
-	"github.com/yohamta/dagu/internal/utils"
+	"github.com/spf13/cobra"
 )
 
-func newStartCommand() *cli.Command {
-	cl := &config.Loader{
-		HomeDir: utils.MustGetUserHomeDir(),
-	}
-	return &cli.Command{
-		Name:  "start",
-		Usage: "dagu start [--params=\"<params>\"] <config>",
-		Flags: []cli.Flag{
-			&cli.StringFlag{
-				Name:     "params",
-				Usage:    "parameters",
-				Value:    "",
-				Required: false,
-			},
-		},
-		Action: func(c *cli.Context) error {
-			config_file_path := c.Args().Get(0)
-			cfg, err := cl.Load(config_file_path, c.String("params"))
-			if err != nil {
-				return err
-			}
-			return start(cfg)
-		},
-	}
+// startCmd represents the start command
+var startCmd = &cobra.Command{
+	Use:   "start",
+	Short: "A brief description of your command",
+	Long: `A longer description that spans multiple lines and likely contains examples
+and usage of using your command. For example:
+
+Cobra is a CLI library for Go that empowers applications.
+This application is a tool to generate the needed files
+to quickly create a Cobra application.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("start called")
+	},
 }
 
-func start(cfg *config.Config) error {
-	a := &dagu.Agent{AgentConfig: &dagu.AgentConfig{
-		DAG: cfg,
-		Dry: false,
-	}}
+func init() {
+	rootCmd.AddCommand(startCmd)
 
-	listenSignals(func(sig os.Signal) {
-		a.Signal(sig)
-	})
+	// Here you will define your flags and configuration settings.
 
-	return a.Run()
+	// Cobra supports Persistent Flags which will work for this command
+	// and all subcommands, e.g.:
+	// startCmd.PersistentFlags().String("foo", "", "A help for foo")
+
+	// Cobra supports local flags which will only run when this command
+	// is called directly, e.g.:
+	// startCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
