+++ cmd/status.go
@@ -1,42 +1,40 @@
-package main
+/*
+Copyright © 2022 NAME HERE <EMAIL ADDRESS>
+
+*/
+package cmd
 
 import (
-	"log"
+	"fmt"
 
-	"github.com/yohamta/dagu/internal/config"
-	"github.com/yohamta/dagu/internal/controller"
-	"github.com/yohamta/dagu/internal/models"
-	"github.com/yohamta/dagu/internal/utils"
-
-	"github.com/urfave/cli/v2"
+	"github.com/spf13/cobra"
 )
 
-func newStatusCommand() *cli.Command {
-	cl := &config.Loader{
-		HomeDir: utils.MustGetUserHomeDir(),
-	}
-	return &cli.Command{
-		Name:  "status",
-		Usage: "dagu status <config>",
-		Action: func(c *cli.Context) error {
-			config_file_path := c.Args().Get(0)
-			cfg, err := cl.Load(config_file_path, "")
-			if err != nil {
-				return err
-			}
-			return queryStatus(cfg)
-		},
-	}
+// statusCmd represents the status command
+var statusCmd = &cobra.Command{
+	Use:   "status",
+	Short: "A brief description of your command",
+	Long: `A longer description that spans multiple lines and likely contains examples
+and usage of using your command. For example:
+
+Cobra is a CLI library for Go that empowers applications.
+This application is a tool to generate the needed files
+to quickly create a Cobra application.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("status called")
+	},
 }
 
-func queryStatus(cfg *config.Config) error {
-	status, err := controller.New(cfg).GetStatus()
-	if err != nil {
-		return err
-	}
-	res := &models.StatusResponse{
-		Status: status,
-	}
-	log.Printf("Pid=%d Status=%s", res.Status.Pid, res.Status.Status)
-	return nil
+func init() {
+	rootCmd.AddCommand(statusCmd)
+
+	// Here you will define your flags and configuration settings.
+
+	// Cobra supports Persistent Flags which will work for this command
+	// and all subcommands, e.g.:
+	// statusCmd.PersistentFlags().String("foo", "", "A help for foo")
+
+	// Cobra supports local flags which will only run when this command
+	// is called directly, e.g.:
+	// statusCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
