+++ cmd/version.go
@@ -1,22 +1,40 @@
-package main
+/*
+Copyright © 2022 NAME HERE <EMAIL ADDRESS>
+
+*/
+package cmd
 
 import (
 	"fmt"
 
-	"github.com/urfave/cli/v2"
-	"github.com/yohamta/dagu/internal/constants"
+	"github.com/spf13/cobra"
 )
 
-func newVersionCommand() *cli.Command {
-	return &cli.Command{
-		Name: "version",
-		Usage: "dagu version",
-		Action: func(c *cli.Context) error {
-			if constants.Version != "" {
-				fmt.Println(constants.Version)
-				return nil
-			}
-			return fmt.Errorf("failed to get version")
-		},
-	}
+// versionCmd represents the version command
+var versionCmd = &cobra.Command{
+	Use:   "version",
+	Short: "A brief description of your command",
+	Long: `A longer description that spans multiple lines and likely contains examples
+and usage of using your command. For example:
+
+Cobra is a CLI library for Go that empowers applications.
+This application is a tool to generate the needed files
+to quickly create a Cobra application.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("version called")
+	},
+}
+
+func init() {
+	rootCmd.AddCommand(versionCmd)
+
+	// Here you will define your flags and configuration settings.
+
+	// Cobra supports Persistent Flags which will work for this command
+	// and all subcommands, e.g.:
+	// versionCmd.PersistentFlags().String("foo", "", "A help for foo")
+
+	// Cobra supports local flags which will only run when this command
+	// is called directly, e.g.:
+	// versionCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
