+++ main.go
@@ -1,0 +1,11 @@
+/*
+Copyright © 2022 NAME HERE <EMAIL ADDRESS>
+
+*/
+package main
+
+import "github.com/yohamta/dagu/cmd"
+
+func main() {
+	cmd.Execute()
+}
